+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 **?.?.? Changes:**
 
-* Added 1 new effect:
+* Added 2 new effects:
   * Activate All Interactables: Interacts with most interactable object on the stage (chests, shrines)
+  * Explosive Deaths: All non-players explode at low health or death, lasts 90 seconds
 
 * Steal Player Items:
   * Reworked steal logic: